Created a GameDIrector who Spawns in the LevelManager when you click on the Screen
LevelManager now spawns in PlatformManager and EffectsManager when it
come stop life and sets relevant attributes (mainly Camera I believe).

In essence control has been given over to the GameDirector as to when a
level actually starts and the LevelManager is a self contained thing
which knows how to run a level of Climber.