Background now uses UV scrolling
Before I had two different images which I was manually moving down
until they were out of view and then I would move them back above the
screen.

Not only was this not as good a method, but it also left little gaps
between images eventually leading to being able to see the clear colour
in the background.

UV scrolling is simpler, cleaner and more correct!