Fixed up the PostGame UI for Tweeting my score
Before I was using the old deprecated OnGUI() method of doing things in
relation to tweeting the score (which was completely borked!), but now
I’m using Canvas elements as I have done with all the rest of the UI.

I took the opportunity to tidy up the GameDirector as well and make it
only Enable/Disable UI when necessary, rather than set individual
values of Text and do bitty work.

Still need to update the Icons as they are all placeholder and not yet
working correctly!