Tweaked InGame Score text size and UV scrolling to keep Platforms inline with Background
Scores can now fit up to 4 digits (eg. 1000) even though I doubt anyone
would get that high, its more of a just incase as it wasn’t able to
display more than 2 digits before (which is clearly wrong as I’ve
gotten over 100 before).

I tweaked a magic number in the UV Scrolling “Background.cs” script,
which was meant to represent the size of the texture (brought it down
from 13.66 to 13.5 as this seems to keep the Platforms more inline with
the scrolling).