Fixed the Incorrect Twitter Msg being posted in HeartPage + Improved TapButton
I was actually posting the same Twitter Message regardless of whether
it was coming from the HeartPage or the PostGame screen, I’ve done a
quick fix for that.

The TapButton still looks garish, but now its a little better as it has
a White finger instead of a black one. One of the final things I’ll
need to realistically do is make this Button a lot nicer.