Updated some Images and Buttons
Changed the Left and Right images that appear at the start of the level
to be Buttons, rather than just images which I hide/show. Its actually
a lot simpler code-wise, although it was a bit of a ball-ache to make
it work correctly, as I had to add it correctly to the prefab (so that
it could call the correct function) and it needed to be a child of the
Canvas Renderer.

Updated the Heart Image as the original looked too ‘gamey’ and looked
odd in the Menu, although now the new one makes Climber himself look
out of place - might change this later.

App Icon has been updated to something a little better than it was
before, it now has Climber with the yellow brick background.

I removed the SplashScreen as I couldn’t get it to look good, I don’t
think its really needed.

Also renamed to project from ClimberUnity, to Climber, so that the App
is named correctly!