Deleted more unnecessary assets + Spotted a bug with TapButton (not yet fixed)
I deleted a bunch more StandardAssets that aren’t being used and some
UnityAds code that I accidentally hadn’t deleted.

I spotted a bug with the TapButton too as they are not always appearing
even though the code is setting the GameObject to be Active correctly.
Ideally I would look into this at some point, but its not a game
breaking bug so I don’t need to worry about it for now.